+++ CHANGELOG.md
@@ -57,6 +57,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Exposed `FilterOptions` so it can be imported from `account::FilterOptions`;
 - Only expose `MessageHandler` methods `SetStrongholdPassword`, `SetStrongholdPasswordClearInterval` and `StoreMnemonic` when `feature = "stronghold"` is enabled; 
 - `Message::{GetAccount, VerifyMnemonic, SetClientOptions, SetStrongholdPassword, SetStrongholdPasswordClearInterval, StoreMnemonic, EmitTestEvent, Bech32ToHex, ClearListeners}` to named fields for better error messages;
+- Made `AccountManager::stop_background_syncing()` async to await until syncing actually stopped;
 - `OutputData::input_signing_data` returns `Result<Option<InputSigningData>>` instead of `Option<InputSigningData>>`;
 
 ### Removed
@@ -65,6 +66,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `IncomingTransactionDataDto` type;
 - `SyncOptions::sync_aliases_and_nfts`;
 - `{TransactionOptions, TransactionOptionsDto}::allow_burning`;
+- Background task spawning to retry a transaction;
 
 ### Fixed